Starting work on new map editor movement system
Starting work on a new movement system for the map editor.
Basically im attempting to copy how the unity editor does movement/scaling and rotating in the future, but I have NO clue how to handle rotating yet.

I also changed the scenes to an enum for future proofing incase the indexes mess up, we dont have to search through the code and change them. Simply just change the index in the GameScenes enum